adbot/backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
oracledb
twilio
python-multipart
